Move make_metric to a static method.
Also moved Metric in its own file
