+++ tests/cli/test_command_graphite_web.py
@@ -29,7 +29,7 @@ class TestCommandGraphiteWeb(bg_test_utils.TestCaseWithFakeAccessor):
         name = 'foo.bar'
         metadata = bg_accessor.MetricMetadata(
             retention=bg_accessor.Retention.from_string('1440*60s'))
-        self.accessor.create_metric(self.make_metric(name, metadata))
+        self.accessor.create_metric(bg_test_utils.make_metric(name, metadata))
         self.accessor.flush()
 
         cmd = command_graphite_web.CommandGraphiteWeb()
