+++ tests/test_accessor.py
@@ -195,14 +195,6 @@ class TestMetricMetadata(unittest.TestCase):
                           "carbon_xfilesfactor", 0.5)
 
 
-class TestMetric(unittest.TestCase):
-
-    def test_dir(self):
-        metric = bg_test_utils.make_metric("a.b.c")
-        self.assertIn("name", dir(metric))
-        self.assertIn("carbon_xfilesfactor", dir(metric))
-
-
 class TestAccessor(bg_test_utils.TestCaseWithFakeAccessor):
 
     def test_context_manager(self):
@@ -238,30 +230,6 @@ class TestAccessor(bg_test_utils.TestCaseWithFakeAccessor):
             self.assertIsNotNone(name)
 
         self.accessor.map(_callback, errback=_errback)
-
-
-class TestAccessorFunctions(unittest.TestCase):
-
-    def test_sanitize_metric_name_should_remove_multiple_dots(self):
-        self.assertEqual("foo.bar.baz", bg_accessor.sanitize_metric_name("foo.bar..baz"))
-        self.assertEqual("foo.bar.baz", bg_accessor.sanitize_metric_name("foo.bar...baz"))
-        self.assertEqual("foo.bar.baz", bg_accessor.sanitize_metric_name("foo..bar..baz"))
-
-    def test_sanitize_metric_name_should_trim_trailing_dots(self):
-        self.assertEqual("foo.bar", bg_accessor.sanitize_metric_name("foo.bar."))
-        self.assertEqual("foo.bar", bg_accessor.sanitize_metric_name("foo.bar.."))
-
-    def test_sanitize_metric_name_should_trim_heading_dots(self):
-        self.assertEqual("foo.bar", bg_accessor.sanitize_metric_name(".foo.bar"))
-        self.assertEqual("foo.bar", bg_accessor.sanitize_metric_name("..foo.bar"))
-
-    def test_sanitize_metric_name_should_handle_None_value(self):
-        self.assertEqual(None, bg_accessor.sanitize_metric_name(None))
-
-    def test_sanitize_metric_name_should_handle_empty_value(self):
-        self.assertEqual("", bg_accessor.sanitize_metric_name(""))
-        self.assertEqual("", bg_accessor.sanitize_metric_name("."))
-        self.assertEqual("", bg_accessor.sanitize_metric_name(".."))
 
 
 class TestPointGrouper(unittest.TestCase):
